stage1_paper/cellpose-inference-gpu/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiobotocore=2.5.4=pypi_0
aiohttp=3.9.1=pypi_0
aioitertools=0.11.0=pypi_0
aiosignal=1.3.1=pypi_0
anndata=0.10.3=pypi_0
annotated-types=0.6.0=pypi_0
array-api-compat=1.4=pypi_0
asciitree=0.3.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
blas=1.0=mkl
botocore=1.31.17=pypi_0
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cellpose=2.2.3=pypi_0
certifi=2023.11.17=py39h06a4308_0
cffi=1.16.0=py39h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
cloudpickle=3.0.0=pypi_0
colorcet=3.0.1=pypi_0
contourpy=1.2.0=pypi_0
cryptography=41.0.7=py39hdda0065_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
dask=2023.12.1=pypi_0
dask-image=2023.8.1=pypi_0
datashader=0.16.0=pypi_0
distributed=2023.12.1=pypi_0
docrep=0.3.2=pypi_0
exceptiongroup=1.2.0=pypi_0
fasteners=0.19=pypi_0
fastremap=1.14.1=pypi_0
ffmpeg=4.3=hf484d3e_0
fiona=1.9.5=pypi_0
fonttools=4.47.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2023.6.0=pypi_0
geopandas=0.14.1=pypi_0
get-annotations=0.1.2=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
h5py=3.10.0=pypi_0
idna=3.4=py39h06a4308_0
igraph=0.10.8=pypi_0
imagecodecs=2023.9.18=pypi_0
imageio=2.33.1=pypi_0
importlib-metadata=7.0.0=pypi_0
importlib-resources=6.1.1=pypi_0
inflect=7.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
leidenalg=0.10.1=pypi_0
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
llvmlite=0.41.1=pypi_0
locket=1.0.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-scalebar=0.8.1=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
msgpack=1.0.7=pypi_0
multidict=6.0.4=pypi_0
multipledispatch=1.0.0=pypi_0
multiscale-spatial-image=0.11.2=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=pypi_0
numba=0.58.1=pypi_0
numcodecs=0.12.1=pypi_0
numpy=1.23.4=pypi_0
ome-zarr=0.8.3=pypi_0
omnipath=1.0.8=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
param=2.0.1=pypi_0
partd=1.4.1=pypi_0
patsy=0.5.4=pypi_0
pillow=10.0.1=py39ha6cbd5a_0
pims=0.6.1=pypi_0
pip=23.3.1=py39h06a4308_0
psutil=5.9.7=pypi_0
pyarrow=14.0.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyct=0.5.0=pypi_0
pydantic=2.5.2=pypi_0
pydantic-core=2.14.5=pypi_0
pygeos=0.14=pypi_0
pygments=2.17.2=pypi_0
pynndescent=0.5.11=pypi_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.1.1=pypi_0
pyproj=3.6.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=1.12.0=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pywavelets=1.5.0=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=py39h06a4308_0
rich=13.7.0=pypi_0
roifile=2023.8.30=pypi_0
s3fs=2023.6.0=pypi_0
scanpy=1.9.6=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.9.1=pypi_0
seaborn=0.12.2=pypi_0
session-info=1.0.0=pypi_0
setuptools=68.2.2=py39h06a4308_0
shapely=2.0.2=pypi_0
six=1.16.0=pypi_0
slicerator=1.1.0=pypi_0
sortedcontainers=2.4.0=pypi_0
spatial-image=0.3.0=pypi_0
spatialdata=0.0.15=pypi_0
sqlite=3.41.2=h5eee18b_0
squidpy=1.3.1=pypi_0
statsmodels=0.14.1=pypi_0
stdlib-list=0.10.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tblib=3.0.0=pypi_0
texttable=1.7.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.12.9=pypi_0
tk=8.6.12=h1ccaba5_0
toolz=0.12.0=pypi_0
torchaudio=0.12.0=py39_cu113
torchvision=0.13.0=py39_cu113
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.3=pypi_0
umap-learn=0.5.5=pypi_0
urllib3=1.26.18=py39h06a4308_0
validators=0.22.0=pypi_0
wheel=0.41.2=py39h06a4308_0
wrapt=1.16.0=pypi_0
xarray=2023.12.0=pypi_0
xarray-dataclasses=1.7.0=pypi_0
xarray-datatree=0.0.13=pypi_0
xarray-schema=0.0.3=pypi_0
xarray-spatial=0.3.7=pypi_0
xz=5.4.5=h5eee18b_0
yarl=1.9.4=pypi_0
zarr=2.16.1=pypi_0
zict=3.0.0=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
